Extract validation from repo.ts
